Segmentation dataset

seg200_sf,seg500_sf,seg800_sf,seg1000_sf: a dataset containing segments generated from PlanetScope image, level 3B, acquired on Feb 18, 2020, with 3.7-meter resolution (Planet Team, 2017), using the multiresolution segmentation method (Baatz and Schape, 2000).

The data covers the approximately the same area of LEM+ dataset (see ref_sf).

The data was post-processed using the spectral difference algorithm on band 3.

The polygons were simplified using the Douglas-Peucker algorithm in QGIS.

Self-intersections were removed using SAGA's Polygon Self-Intersection.

Segmentation parameters:

scale parameter: 200 (seg200_sf), 500 (seg500_sf), 800 (seg800_sf), and 1000 (seg1000_sf)

shape: 0.9

compactness: 0.1

Spectral difference parameters:

spectral difference: 20

Simplification parameter:

distance: 10-meters

Only those polygons intersecting reference data with an area-perimeter ratio above 25 were selected.

sample_seg_sf: a subset of seg_sf dataset.

Usage

seg200_sf

seg500_sf

seg800_sf

seg1000_sf

sample_seg_sf

Format

These datasets are objects of class sf (inherited from tbl_df, tbl, data.frame) with 2 variables:

id: identification

geometry: polygons

seg200_sf: a dataset with 547 features. seg500_sf: a dataset with 215 features. seg800_sf: a dataset with 169 features. seg1000_sf: a dataset with 158 features.

An object of class sf (inherits from tbl_df, tbl, data.frame) with 215 rows and 2 columns.

An object of class sf (inherits from tbl_df, tbl, data.frame) with 169 rows and 2 columns.

An object of class sf (inherits from tbl_df, tbl, data.frame) with 158 rows and 2 columns.

sample_seg_sf: a dataset with 6 features extracted from seg500_sf dataset.

References

Planet Team, 2017. Planet Application Program Interface: In Space for Life on Earth. San Francisco, CA. https://www.planet.com

Baatz, M., Schape, A., 2000. Multiresolution segmentation - an optimization approach for high quality multi-scale image segmentation. In: Strobl, J., Blaschke, T., Griesebner, G. (Eds.), Angewandte Geographische Informations-Verarbeitung XII. Wichmann Verlag, Karlsruhe, Germany, pp. 12-23. <>

Examples

data("seg200_sf", package = "segmetric")
data("seg500_sf", package = "segmetric")
data("seg800_sf", package = "segmetric")
data("seg1000_sf", package = "segmetric")
data("sample_seg_sf", package = "segmetric")