Set functions

These functions compute subsets required to calculate segmentation metrics as described in Clinton et al. (2010) and Costa et al. (2017).

sm_ref() returns the set of \(n\) polygons of reference, represented by \(X = \{x_{i}: i = 1, ....., n\}\)

sm_seg() returns the set of \(m\) segmentation polygons, represented by \(Y = \{y_{j}: j = 1, ....., m\}\)

sm_ytilde() returns \(\tilde{Y}_{i}\), a subset of \(Y\), where \(\tilde{Y}_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) \neq 0\}\)

sm_xtilde() returns \(\tilde{X}_{j}\), a subset of \(X\), where \(\tilde{X}_{j} = \{x_{i}: \rm{area}(y_{j} \cap x_{i}) \neq 0\}\)

sm_yprime() returns \(Y'_{i}\), a subset of \(Y\), where \(Y'_{i} = \{y_{j}: max(\rm{area}(x_{i} \cap y_{j}))\}\)

sm_xprime() returns \(X'_{j}\), a subset of \(X\), where \(X'_{j} = \{x_{i}: max(\rm{area}(y_{j} \cap x_{i}))\}\)

sm_ya() returns \(Y\!a_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!a_{i} = \{y_{j}: \rm{centroid}(x_{i}) \:\rm{in}\: y_{j}\}\)

sm_yb() returns \(Y\!b_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!b_{i} = \{y_{j}: \rm{centroid}(y_{j}) \:\rm{in}\: x_{i}\}\)

sm_yc() returns \(Y\!c_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!c_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) / \rm{area}(y_{j}) > 0.5\}\)

sm_yd() returns \(Y\!d_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!d_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) / \rm{area}(x_{i}) > 0.5\}\)

sm_ystar() returns \({Y}^{*}_{i}\), where \({Y}^{*}_{i} = Y\!a_{i} \cup Y\!b_{i} \cup Y\!c_{i} \cup Y\!c_{i}\)

sm_ycd() returns \(Y\!cd_{i}\), where \(Y\!cd_{i} = Y\!c_{i} \cup Y\!d_{i}\)

sm_ye() returns \(Y\!e_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!e_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) / \rm{area}(y_{j}) = 1\}\)

sm_yf() returns \(Y\!f_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!f_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) / \rm{area}(y_{j}) > 0.55\}\)

sm_yg() returns \(Y\!g_{i}\), a subset of \(\tilde{Y}_{i}\), where \(Y\!g_{i} = \{y_{j}: \rm{area}(x_{i} \cap y_{j}) / \rm{area}(y_{j}) > 0.75\}\)

Usage

sm_ytilde(m)

sm_xtilde(m)

sm_yprime(m)

sm_xprime(m)

sm_ya(m)

sm_yb(m)

sm_yc(m)

sm_yd(m)

sm_ystar(m)

sm_ycd(m)

sm_ye(m)

sm_yf(m)

sm_yg(m)

Arguments

m

A segmetric object.

Value

sm_ref(): Return an object of class ref_sf (inherited from sf) containing identification (ref_id) and geometry (geometry) columns.

sm_seg(): Return an object of class seg_sf (inherited from sf) containing identification (seg_id) and geometry (geometry) columns.

sm_ytilde(), sm_xtilde(), sm_yprime(), sm_xprime(), sm_ya(), sm_yb(), sm_yc(), sm_yd(), sm_ystar(), sm_ycd(), sm_ye(), sm_yf(), and sm_yg(): Return an object of class subset_sf (inherited from sf) containing identification (ref_id and seg_id), and geometry (geometry) columns.

References

Clinton, N., Holt, A., Scarborough, J., Yan, L., & Gong, P. (2010). Accuracy Assessment Measures for Object-based Image Segmentation Goodness. Photogrammetric Engineering & Remote Sensing, 76(3), 289–299. doi:10.14358/PERS.76.3.289 .

Costa, H., Foody, G. M., & Boyd, D. S. (2018). Supervised methods of image segmentation accuracy assessment in land cover mapping. Remote Sensing of Environment, 205(December 2017), 338–351. doi:10.1016/j.rse.2017.11.024 .